v2: logic separation into userPlayers hook
